% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19}
\alias{covid19}
\title{COVID-19 Cases Time Series in India}
\format{
An object of class \code{data.frame} with 236 rows and 7 columns.

\describe{
\item{Date}{Date as a character string}
\item{Daily.Confirmed}{Daily confirmed cases as an integer vector}
\item{Total.Confirmed}{Total confirmed cases upto current date as an integer vector}
\item{Daily.Recovered}{Daily recovered cases an an integer vector}
\item{Total.Recovered}{Total confirmed cases upto current date as an integer vector}
\item{Daily.Deceased}{Daily deceased cases as an integer vector}
\item{Total.Deceased}{Total deceased cases upto current date as an integer vector}
}
}
\usage{
covid19
}
\description{
Contains dailies and totals of cases, recoveries, and deaths from the COVID-19
outbreak in India from January 30 to September 21 of 2020.
}
\examples{
covid19
head(covid19)
tail(covid19)
}
\keyword{datasets}
